refactor: refactor trivial points according to the review related to #10
